refactor: cleanup and rename hooks
