+++ packages/app/src/hooks/index.js
@@ -1,0 +1,62 @@
+import { useContext, useState, useEffect } from 'react';
+import jwt from 'jsonwebtoken';
+import AuthContext from '../context/AuthContext';
+import { BASE_URL } from '../services/util';
+
+const useTokenOrRefresh = () => {
+    const authContext = useContext(AuthContext);
+    const [token, setToken] = useState(authContext.token);
+    const tokenExpires = jwt.decode(token).exp;
+    const now = Date.now().valueOf() / 1000;
+
+    useEffect(() => {
+        const getNewToken = async () => {
+            const res = await fetch(`${BASE_URL}/api/auth/refreshtoken`, {
+                method: 'GET',
+                credentials: 'include'
+            });
+            const data = await res.json();
+            if (data && data.payload) {
+                setToken(data.payload.accessToken);
+                authContext.updateToken(data.payload.accessToken);
+            }
+        };
+        if (now > tokenExpires) {
+            getNewToken();
+        }
+        // eslint-disable-next-line react-hooks/exhaustive-deps
+    }, []);
+    return token;
+};
+
+const useFetchData = (endpoint, options = {}) => {
+    const [response, setResponse] = useState(null);
+    const [error, setError] = useState(null);
+    const [isFetching, setIsFetching] = useState(true);
+    const token = useTokenOrRefresh();
+
+    useEffect(() => {
+        const fetchData = async () => {
+            try {
+                const res = await fetch(`${BASE_URL}${endpoint}`, {
+                    ...options,
+                    method: 'GET',
+                    credentials: 'include',
+                    headers: { Authorization: `Bearer ${token}` }
+                });
+                const json = await res.json();
+                setResponse(json);
+                setIsFetching(false);
+            } catch (error) {
+                console.log(error);
+                setError(error);
+                setIsFetching(false);
+            }
+        };
+        fetchData();
+        // eslint-disable-next-line react-hooks/exhaustive-deps
+    }, [token]);
+
+    return { response, error, isFetching };
+};
+export { useFetchData };
